refactor(core): Extract file processing logic to separate modules
- Create fileRead.ts for file reading functionality

- Create fileProcessContent.ts for content processing

- Update worker files to use new modules
