+++ src/core/file/workers/fileCollectWorker.ts
@@ -1,10 +1,6 @@
-import * as fs from 'node:fs/promises';
 import path from 'node:path';
-import iconv from 'iconv-lite';
-import { isBinary } from 'istextorbinary';
-import jschardet from 'jschardet';
-import pc from 'picocolors';
 import { logger, setLogLevelByEnv } from '../../../shared/logger.js';
+import { readRawFile } from '../fileRead.js';
 
 export interface FileCollectTask {
   filePath: string;
@@ -27,39 +23,4 @@ export default async ({ filePath, rootDir, maxFileSize }: FileCollectTask) => {
   }
 
   return null;
-};
-
-const readRawFile = async (filePath: string, maxFileSize: number): Promise<string | null> => {
-  try {
-    const stats = await fs.stat(filePath);
-
-    if (stats.size > maxFileSize) {
-      const sizeKB = (stats.size / 1024).toFixed(1);
-      const maxSizeKB = (maxFileSize / 1024).toFixed(1);
-      logger.trace(`File exceeds size limit: ${sizeKB}KB > ${maxSizeKB}KB (${filePath})`);
-      return null;
-    }
-
-    if (isBinary(filePath)) {
-      logger.debug(`Skipping binary file: ${filePath}`);
-      return null;
-    }
-
-    logger.trace(`Reading file: ${filePath}`);
-
-    const buffer = await fs.readFile(filePath);
-
-    if (isBinary(null, buffer)) {
-      logger.debug(`Skipping binary file (content check): ${filePath}`);
-      return null;
-    }
-
-    const encoding = jschardet.detect(buffer).encoding || 'utf-8';
-    const content = iconv.decode(buffer, encoding);
-
-    return content;
-  } catch (error) {
-    logger.warn(`Failed to read file: ${filePath}`, error);
-    return null;
-  }
 };
